// modules/apps/headless/headless-admin-site/headless-admin-site-client-js/src/models/FragmentMappedValue.ts
/**
 * SPDX-FileCopyrightText: (c) 2026 Liferay, Inc. https://liferay.com
 * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
 */


/**
 * @author Rubén Pulido
 * @generated
 */

	/**
	* A fragment mapped value.
	*/
	export class FragmentMappedValue {
			"mapping"?: object;

		static "discriminator": string | undefined = undefined;

	static "attributeTypeMap": Array<{
		baseName: string;
		name: string;
		type: string;
	}> = [
		{
			baseName: "mapping",
			name: "mapping",
			type: "object",
		},
		];

		static getAttributeTypeMap() {
				return FragmentMappedValue.attributeTypeMap;
		}
	}